Fixes packages, sources folders and Implicits name
